requirements.txt
streamlit
google-generative-ai
PyPDF2
python-docx
Pillow
python-dotenv
